Locate server ressources at /api for production
To ensure URL consistency between production and development and
to avoid conflicting URLs between client and server projects.
