+++ packages/server/src/index.ts
@@ -2,33 +2,15 @@ import express from "express";
 import session from "express-session";
 import http from "http";
 import { Server } from "socket.io";
-import {
-  getGame,
-  getGames,
-  createGame,
-  playMove,
-  takeSeat,
-  leaveSeat,
-} from "./games";
-import {
-  createUserWithSessionId,
-  deleteUser,
-  getUser,
-  getUserBySessionId,
-} from "./users";
+import { createUserWithSessionId, getUser, getUserBySessionId } from "./users";
 import bodyParser from "body-parser";
 import cors from "cors";
 import path from "path";
 import { connectToDb } from "./db";
 import passport from "passport";
 import { Strategy as CustomStrategy } from "passport-custom";
-import {
-  GameResponse,
-  MovesType,
-  DELETETHIS_getCurrentUser,
-  User,
-  UserResponse,
-} from "@ogfcommunity/variants-shared";
+import { UserResponse } from "@ogfcommunity/variants-shared";
+import { router as apiRouter } from "./api";
 
 const LOCAL_ORIGIN = "http://localhost:3000";
 
@@ -122,107 +104,7 @@ const io = new Server(server, {
 // Initialize MongoDB
 connectToDb();
 
-// Set up express routes
-app.get("/games/:gameId", async (req, res) => {
-  const game: GameResponse = await getGame(req.params.gameId);
-  res.send(game);
-});
-
-app.get("/games", async (req, res) => {
-  const games: GameResponse[] = await getGames(Number(req.query.page));
-  res.send(games || 0);
-});
-
-app.post("/games", async (req, res) => {
-  const data = req.body;
-
-  const game: GameResponse = await createGame(data.variant, data.config);
-
-  res.send(game);
-
-  return;
-});
-
-app.post("/games/:gameId/move", async (req, res, next) => {
-  const move: MovesType = req.body;
-  // TODO: make sure this is set to a valid id once we have user auth
-  // const user_id = req.user?.id; */
-  const user: User = DELETETHIS_getCurrentUser();
-
-  try {
-    res.send(await playMove(req.params.gameId, move, user.id));
-  } catch (e) {
-    next(e.message);
-  }
-});
-
-app.post("/games/:gameId/sit/:seat", async (req, res) => {
-  const move: MovesType = req.body;
-  // TODO: make sure this is set to a valid id once we have user auth
-  // const user_id = req.user?.id; */
-  const user: User = DELETETHIS_getCurrentUser();
-
-  const players: User[] = await takeSeat(
-    req.params.gameId,
-    Number(req.params.seat),
-    user
-  );
-
-  res.send(players);
-});
-
-app.post("/games/:gameId/leave/:seat", async (req, res) => {
-  const move: MovesType = req.body;
-  // TODO: make sure this is set to a valid id once we have user auth
-  // const user_id = req.user?.id; */
-  const user = DELETETHIS_getCurrentUser();
-
-  const players: User[] = await leaveSeat(
-    req.params.gameId,
-    Number(req.params.seat),
-    user.id
-  );
-
-  res.send(players);
-});
-
-app.get("/guestLogin", function (req, res, next) {
-  passport.authenticate("guest", (err, user, info) => {
-    if (err) {
-      return next(err);
-    }
-    if (!user) {
-      return next(new Error(info.message));
-    }
-
-    req.logIn(user, function (err) {
-      if (err) {
-        return next(err);
-      }
-
-      return res.json(user);
-    });
-  })(req, res, next);
-});
-
-app.get("/checkLogin", function (req, res, next) {
-  return res.json(req.user ? req.user : null);
-});
-
-app.get("/logout", async function (req, res) {
-  const user = req.user as UserResponse;
-  if (user.login_type === "guest") {
-    deleteUser(user.id);
-  }
-  req.logout((err) => {
-    if (err) throw new Error(err);
-    req.session.destroy((err) => {
-      if (err) throw new Error(err);
-      // res.sendStatus(200); // TODO: client only expects JSON
-      res.json({});
-    });
-  });
-});
+app.use("/api", apiRouter);
 
 io.on("connection", (socket) => {
   console.log("a user connected");
