+++ packages/vue-client/vite.config.ts
@@ -18,7 +18,6 @@ export default defineConfig({
         secure: false,
         target: "http://localhost:3001/",
         changeOrigin: true,
-        rewrite: (path) => path.replace(/^\/api/, ""),
       },
     },
   },
